Finalize command, messages, vault support.

+++ eternaleconomy-core/src/main/java/com/eternalcode/economy/command/admin/AdminRemoveCommand.java
@@ -3,27 +3,74 @@ package com.eternalcode.economy.command.admin;
 import com.eternalcode.economy.EconomyPermissionConstant;
 import com.eternalcode.economy.account.Account;
 import com.eternalcode.economy.account.AccountPaymentService;
+import com.eternalcode.economy.format.DecimalFormatter;
+import com.eternalcode.economy.multification.NoticeService;
 import dev.rollczi.litecommands.annotations.argument.Arg;
 import dev.rollczi.litecommands.annotations.command.Command;
 import dev.rollczi.litecommands.annotations.context.Context;
 import dev.rollczi.litecommands.annotations.execute.Execute;
 import dev.rollczi.litecommands.annotations.permission.Permission;
-import org.bukkit.command.CommandSender;
-
 import java.math.BigDecimal;
+import org.bukkit.command.CommandSender;
 
 @Command(name = "economy admin remove")
 @Permission(EconomyPermissionConstant.ADMIN_REMOVE_PERMISSION)
 public class AdminRemoveCommand {
 
     private final AccountPaymentService accountPaymentService;
+    private final DecimalFormatter decimalFormatter;
+    private final NoticeService noticeService;
 
-    public AdminRemoveCommand(AccountPaymentService accountPaymentService) {
+    public AdminRemoveCommand(
+            AccountPaymentService accountPaymentService,
+            DecimalFormatter decimalFormatter,
+            NoticeService noticeService
+    ) {
         this.accountPaymentService = accountPaymentService;
+        this.decimalFormatter = decimalFormatter;
+        this.noticeService = noticeService;
     }
 
     @Execute
     void execute(@Context CommandSender sender, @Arg Account receiver, @Arg BigDecimal amount) {
-        this.accountPaymentService.removeBalance(receiver, amount);
+        if (amount.compareTo(BigDecimal.ZERO) < 0) {
+            this.noticeService.create()
+                    .notice(notice -> notice.invalidAmount)
+                    .placeholder("{AMOUNT}", amount.toString())
+                    .viewer(sender)
+                    .send();
+
+            return;
+        }
+
+        if (receiver.balance().compareTo(amount) < 0) {
+            BigDecimal subtract = amount.subtract(receiver.balance());
+            this.noticeService.create()
+                    .notice(notice -> notice.admin.insufficientFunds)
+                    .placeholder("{PLAYER}", receiver.name())
+                    .placeholder("{MISSING_BALANCE}", this.decimalFormatter.format(subtract))
+                    .viewer(sender)
+                    .send();
+
+            return;
+        }
+
+        boolean successful = this.accountPaymentService.removeBalance(receiver, amount);
+
+        if (successful) {
+            this.noticeService.create()
+                    .notice(notice -> notice.admin.removed)
+                    .placeholder("{AMOUNT}", this.decimalFormatter.format(amount))
+                    .placeholder("{PLAYER}", receiver.name())
+                    .viewer(sender)
+                    .send();
+
+            this.noticeService.create()
+                    .notice(notice -> notice.player.removed)
+                    .placeholder("{AMOUNT}", this.decimalFormatter.format(amount))
+                    .placeholder("{PLAYER}", receiver.name())
+                    .player(receiver.uuid())
+                    .send();
+        }
     }
 }
